Move profile media methods to media processor
