+++ internal/processing/processor.go
@@ -174,12 +174,13 @@ func NewProcessor(
 	// Start with sub processors that will
 	// be required by the workers processor.
 	common := common.New(state, converter, federator, filter)
+	processor.account = account.New(&common, state, converter, mediaManager, oauthServer, federator, filter, parseMentionFunc)
 	processor.media = media.New(state, converter, mediaManager, federator.TransportController())
-	processor.account = account.New(&common, state, converter, &processor.media, mediaManager, oauthServer, federator, filter, parseMentionFunc)
 	processor.stream = stream.New(state, oauthServer)
 
 	// Instantiate the rest of the sub
 	// processors + pin them to this struct.
+	processor.account = account.New(&common, state, converter, mediaManager, oauthServer, federator, filter, parseMentionFunc)
 	processor.admin = admin.New(state, cleaner, converter, mediaManager, federator.TransportController(), emailSender)
 	processor.fedi = fedi.New(state, &common, converter, federator, filter)
 	processor.filtersv1 = filtersv1.New(state, converter)
